examples/servers/smack/pyproject.toml
[project]
name = "mcp-smack"
version = "0.1.0"
description = "SMACK application"
requires-python = ">=3.10"
authors = [{ name = "Featureform, Inc." }]
keywords = ["mcpengine", "mcp", "llm", "automation", "web", "fetch"]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
dependencies = ["psycopg2-binary==2.9.9", "mcpengine[cli,lambda]>=0.2.0"]

[project.scripts]
mcp-smack = "mcp_smack.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_smack"]

[tool.pyright]
include = ["mcp_smack"]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.uv]
dev-dependencies = ["pyright>=1.1.378", "pytest>=8.3.3", "ruff>=0.6.9"]
